implementada a função de alterar cliente

+++ src/tela/ler.rs
@@ -6,7 +6,7 @@ pub fn ler_dados() -> String {
     entrada.trim().to_string()
 }
 
-pub fn ler_dados_int() -> i32 {
+pub fn ler_dados_int() -> usize {
     let mut entrada = String::new();
     io::stdin().read_line(&mut entrada).expect("Falha ao ler a entrada");
     entrada.trim().parse().expect("Falha ao converter para inteiro")
